+++ frontend/src/hooks/channel/index.ts
@@ -1,2 +1,3 @@
 export * from './useChannel';
 export * from './useLeaveChannel';
+export * from './useChat';
