+++ frontend/src/hooks/channel/useChat.tsx
@@ -2,12 +2,12 @@ import { emitEvent } from '@/libs/api';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { channelDataState, channelIdState } from '@/stores';
 import { SendChat } from '@/dto/channel/socket';
-import { useUserInfo } from '../user';
+import { useAuth } from '../auth';
 
 export const useChat = () => {
   const channelId = useRecoilValue(channelIdState);
   const setNewChannelData = useSetRecoilState(channelDataState);
-  const { userInfo } = useUserInfo();
+  const { userInfo } = useAuth();
 
   const sendChat = (content: string) => {
     const data: SendChat = {
