+++ frontend/src/hooks/user/index.ts
@@ -2,4 +2,3 @@ export * from './useFileUpload';
 export * from './useHistroyData';
 export * from './usePatchNickName';
 export * from './useProfileData';
-export * from './useUserInfo';
