+++ frontend/src/libs/utils/itemgenerator.tsx
@@ -1,9 +1,13 @@
-import { useChannelMutation, useLeaveChannel } from '@/hooks/channel';
-import { useFriendMutation } from '@/hooks/friend';
-import { useBlockedMutation, useBlocked } from '@/hooks/blocked';
+import {
+  useChannelMutation,
+  useLeaveChannel,
+  useFriendMutation,
+  useBlockedMutation,
+  useBlocked,
+  useAuth,
+} from '@/hooks';
 import { useRecoilValue } from 'recoil';
 import { channelIdState, channelDataState, currentRoleSelector } from '@/stores';
-import { useUserInfo } from '@/hooks/user';
 import { MemberInfo } from '@/dto/channel';
 import { useNavigate } from 'react-router-dom';
 
@@ -27,7 +31,7 @@ export const useItemGenerator = (): Items => {
   const newChannelData = useRecoilValue(channelDataState);
   const currentRole = useRecoilValue(currentRoleSelector);
   const channelId = useRecoilValue(channelIdState);
-  const { userInfo } = useUserInfo();
+  const { userInfo } = useAuth();
   const { refetchBlocked } = useBlocked();
   const navigate = useNavigate();
