+++ frontend/src/pages/GamePage/GameResultModal/GameResultModal.tsx
@@ -11,10 +11,11 @@ import { useRecoilValue, useSetRecoilState } from 'recoil';
 import { MemberInfo } from '@/dto/channel/socket';
 import { Grid, WideModal, Text, Avatar, GameButton } from '@/common';
 import { useNavigate } from 'react-router-dom';
-import { useLeaveChannel } from '@/hooks/channel';
+import { useLeaveChannel, useAuth } from '@/hooks';
 
 export const GameResultModal = ({ isEnd }: { isEnd: boolean }) => {
   const navigate = useNavigate();
+  const { refetch: refetchAuth } = useAuth();
   const setChannelData = useSetRecoilState(channelDataState);
   const setGameStatus = useSetRecoilState(gameStatusState);
 
@@ -37,6 +38,13 @@ export const GameResultModal = ({ isEnd }: { isEnd: boolean }) => {
     image: '',
     role: 'admin',
   });
+
+  useEffect(() => {
+    refetchAuth();
+    return () => {
+      refetchAuth();
+    };
+  }, []);
 
   useEffect(() => {
     if (gameResult && gamePlayer.leftPlayer && gamePlayer.rightPlayer) {
