+++ frontend/src/pages/GameReadyPage/GameReadyPage.tsx
@@ -5,15 +5,13 @@ import { ObserverBox } from './ObserverBox';
 import { Grid, GameButton, Text } from '@/common';
 import { useResetRecoilState, useSetRecoilState, useRecoilState, useRecoilCallback } from 'recoil';
 import { channelIdState, channelDataState, socketState, gameModeState, gameStatusState, blockedIdList } from '@/stores';
-import { useGameMutation, useGameStart } from '@/hooks/game';
+import { useGameMutation, useGameStart, useChannel, useLeaveChannel, useBlocked } from '@/hooks';
 import { Dropdown } from '@/common/Dropdown';
 import { PingPongGame } from '../GamePage';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { useChannel, useLeaveChannel } from '@/hooks/channel';
 import { ChangeEvent, useEffect, useState } from 'react';
 import { Items, useItemGenerator } from '@/libs/utils/itemgenerator';
 import { GameSideBar } from '@/layout/GameLayout/GameSideBar';
-import { useBlocked } from '@/hooks/blocked';
 import { offEvent, onEvent } from '@/libs/api';
 import { ChannelEvent } from '@/constants';
 import { User } from '@/types';
