+++ frontend/src/pages/MessagePage/FriendsModal/FriendsModal.tsx
@@ -1,6 +1,5 @@
 import { Avatar, Box, CommonButton, GameButton, GameInput, Grid, Text, Modal, ModalProps } from '@/common';
-import { useInput } from '@/hooks';
-import { useFriendMutation, useFriendRequest } from '@/hooks/friend';
+import { useInput, useFriendMutation, useFriendRequest } from '@/hooks';
 import { User } from '@/types/entity';
 import { useEffect } from 'react';
