+++ frontend/src/pages/MessagePage/Message/MessageContent.tsx
@@ -1,8 +1,6 @@
 import { Avatar, Box, Grid, Text } from '@/common';
 import { Fragment, useEffect, useRef } from 'react';
-import { useUserInfo } from '@/hooks/user';
-import { useMessages, useNewMessages } from '@/hooks/message';
-import { useIntersectObserver } from '@/hooks/useIntersectObserver';
+import { useAuth, useMessages, useNewMessages, useIntersectObserver } from '@/hooks';
 import { Message, User } from '@/types/entity';
 import { formatTime } from '@/libs/utils';
 import { nanoid } from 'nanoid';
@@ -31,7 +29,7 @@ export const MessageContentItem = ({ side, user, content, createdAt = '' }: Mess
 };
 
 export const MessageContent = () => {
-  const { userInfo } = useUserInfo();
+  const { userInfo } = useAuth();
   const { currentFriend, newMessages } = useNewMessages();
   const {
     messages: { pages },
