+++ frontend/src/pages/MessagePage/Message/MessageInput.tsx
@@ -1,8 +1,7 @@
 import styled from 'styled-components';
 import { ReactComponent as SendSvg } from '@/svgs/send.svg';
 import { Box, IconButton } from '@/common';
-import { useInput } from '@/hooks';
-import { useNewMessages } from '@/hooks/message/useNewMessages';
+import { useInput, useNewMessages } from '@/hooks';
 import { useState } from 'react';
 
 const StyledInput = styled.input`
